# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BilibiliComment Admin API
  version: 0.1.0
  description: 单管理员任务管理、评论观测与系统状态 API。所有 bigint ID 均按字符串传输。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      security: []
      responses:
        '200':
          description: 当前管理员
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUser'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: login
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUser'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        '204': {description: 已退出登录}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /api/v1/tasks/{taskId}/comments:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    get:
      tags: [Executions]
      operationId: listTaskComments
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: 当前任务独立评论表的游标分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
  /api/v1/tasks/{taskId}/comments/search:
    parameters:
      - {$ref: '#/components/parameters/TaskId'}
    post:
      tags: [Comments]
      operationId: searchTaskComments
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CommentSearchRequest'}
      responses:
        '200':
          description: 固定快照内的筛选评论分页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CommentSearchPage'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '410': {$ref: '#/components/responses/Gone'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/RateLimited'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema: {$ref: '#/components/schemas/PositiveId'}
...
    Cursor:
      name: cursor
      in: query
      description: 服务端返回的不透明分页边界；客户端必须原样回传，不得解析或自行构造。
      schema: {type: string, minLength: 1}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
...
  responses:
    BadRequest:
      description: 请求格式错误
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Unauthorized:
      description: 未登录或凭据错误
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Forbidden:
      description: CSRF 失效或来源域名被禁止
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    NotFound:
      description: 资源不存在
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
...
    Gone:
      description: 导出任务或文件已过期并被清理
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    Unprocessable:
      description: 参数验证失败
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
    RateLimited:
      description: 请求过于频繁
      headers:
        Retry-After: {schema: {type: integer, minimum: 1}}
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/ApiProblem'}}}
...
  schemas:
    PositiveId: {type: string, pattern: '^[1-9][0-9]*$'}
...
    CommentReplyScope: {type: string, enum: [ALL, ROOT, REPLIES], default: ALL}
    CommentSort: {type: string, enum: [CTIME_DESC, CTIME_ASC], default: CTIME_DESC}
...
    AdminUser:
      type: object
      required: [username, displayName]
      properties:
        username: {type: string}
        displayName: {type: string}
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 128}
        password: {type: string, minLength: 1, maxLength: 512, format: password, writeOnly: true}
...
    Comment:
      type: object
      required: [id, rpid, parentRpid, mid, uname, avatar, currentLevel, content, ctime]
      properties:
        id: {$ref: '#/components/schemas/PositiveId'}
        rpid: {$ref: '#/components/schemas/PositiveId'}
        parentRpid: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}
        mid: {type: string}
        uname: {type: [string, 'null']}
        avatar: {type: [string, 'null']}
        currentLevel: {type: [integer, 'null']}
        content: {type: [string, 'null']}
        ctime: {type: string, format: date-time}
    CommentFilter:
      type: object
      properties:
        keyword: {type: [string, 'null'], maxLength: 200}
        mid: {type: [string, 'null'], pattern: '^[1-9][0-9]{0,31}$'}
        uname: {type: [string, 'null'], maxLength: 255}
        levelMin: {type: [integer, 'null'], minimum: 0, maximum: 6}
        levelMax: {type: [integer, 'null'], minimum: 0, maximum: 6}
        unknownLevelOnly: {type: boolean, default: false}
        ctimeFrom: {type: [string, 'null'], format: date-time}
        ctimeBefore: {type: [string, 'null'], format: date-time}
        rpid: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}
        parentRpid: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}
        replyScope: {$ref: '#/components/schemas/CommentReplyScope'}
    CommentSearchRequest:
      type: object
      required: [filter]
      properties:
        filter: {$ref: '#/components/schemas/CommentFilter'}
        sort: {$ref: '#/components/schemas/CommentSort'}
        cursor: {type: [string, 'null'], maxLength: 2048}
        limit: {type: integer, minimum: 1, maximum: 100, default: 50}
        includeTotal: {type: boolean, default: false}
    CommentSearchPage:
      type: object
      required: [items, nextCursor, total]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Comment'}
        nextCursor: {type: [string, 'null']}
        total: {type: [integer, 'null'], minimum: 0}
...
    CommentPage:
      type: object
      required: [items, nextCursor, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Comment'}}
        nextCursor: {type: [string, 'null']}
        total: {type: integer, minimum: 0}
    ApiProblem:
      type: object
      required: [status, code, detail]
      properties:
        type: {type: [string, 'null'], format: uri}
        title: {type: [string, 'null']}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string}
        detail: {type: string}
        traceId: {type: [string, 'null']}
        fieldErrors: {type: object, additionalProperties: {type: string}}
        existingTaskId: {oneOf: [{$ref: '#/components/schemas/PositiveId'}, {type: 'null'}]}